requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
hypercorn>=0.14.4
openai>=1.0.0
supabase>=1.2.0
langchain-text-splitters>=0.0.1
langchain-community>=0.0.13
PyPDF2>=3.0.0
python-multipart>=0.0.6
chromadb>=0.4.15
aiohttp>=3.8.5
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
python-pptx>=0.6.21
sentence-transformers>=2.2.2
torch>=2.0.0
pydantic>=2.4.0
cachetools>=5.3.1
psutil>=5.9.0
python-dotenv>=1.0.0
httpx>=0.24.1
groq>=0.4.0
asyncio>=3.4.3
python-docx>=0.8.11
rank_bm25>=0.2.2
pinecone>=6.0.0
youtube_transcript_api>=0.6.2
boto3>=1.28.0
cohere>=4.37
pytube
pypdf
unstructured
requests>=2.31.0
tiktoken
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
cryptography>=41.0.0
